# Move Liveblog block styles into the head