--- a/static/src/stylesheets/module/content/_tweets.scss
+++ b/static/src/stylesheets/module/content/_tweets.scss
@@ -35,7 +35,7 @@
         // on upgrade
         max-width: 500px;
 
-        margin-right: $block-padding-right;
+        margin-right: $gs-gutter;
     }
 }