--- a/static/src/stylesheets/module/content/live-blog/_live-blog.head.scss
+++ b/static/src/stylesheets/module/content/live-blog/_live-blog.head.scss
@@ -373,7 +373,8 @@ $block-padding-left: gs-span(1) + $gs-gutter;
         > .element-comment,
         > .element-embed,
         > .element-audio,
-        > .element-interactive {
+        > .element-interactive,
+        > .element-atom:not(.element-atom--media) {
             width: auto;
             margin: $gs-baseline $gs-gutter/2 ($gs-baseline / 3) * 4;