… static/src/stylesheets/module/content/_live-blog.scss
… static/src/stylesheets/module/content/live-blog/navigation.scss