## What does this change?

Styles for below the fold Liveblog elements (such as blocks) were previously loaded in asynchronously in content.css. Users can however follow links directly to below the fold blocks, and if they do so they'd see a significant jump as the styles are applied to blocks. This branch moves the liveblog block element styles into the head so they are applied earlier. There's still a jump but it's an improvement.   

## What is the value of this and can you measure success?

Styles applied earlier to liveblog blocks, better user experience.

## Does this affect other platforms - Amp, Apps, etc?

No

## Screenshots

Before...

![blog-before](https://user-images.githubusercontent.com/1590704/27950258-8b805bb8-62f8-11e7-9689-0c5f52e2fb3a.gif)

After...

![blog-after](https://user-images.githubusercontent.com/1590704/27950273-94fd7022-62f8-11e7-8b06-cef44fafb884.gif)

## Tested in CODE?

Yes