@sndrs 2017-07-10
do you know how much weight this adds to the head? 

my gut says we have to accept some things require the second CSS load (mainly below the fold), and i wonder how often these blocks are actually seen by people to repaint (i.e. count as above the fold)?