@GHaberis 2017-07-10
@sndrs I'll get this on CODE today to get some numbers out, hopefully I can get enough out to let us weigh up the costs/benefits.