@GHaberis 2017-07-25
@sndrs I revisited this and ran a comparison on PROD and CODE on the same liveblog at the same point in time and the headline stats are below.

**PROD**
First meaningful paint from lighthouse: 1,080 ms 
Consistently Interactive from lighthouse: 11,780 ms
Document weight from dev tools: 89.5 KB

**CODE**
First meaningful paint from lighthouse: 1,060 ms
Consistently Interactive from lighthouse: 9,910 ms
Document weight from dev tools: 97.5 KB

** Document weight on CODE consistently appears to be 6K more than PROD even when comparing like-for-like builds.