@gustavpursche 2017-08-30
@GHaberis Is this something you still work on?