@GHaberis 2017-08-30
@gustavpursche I was actually discussing with @SiAdcock yesterday whether to close this. We decided I'd get some comparisons off Code (using master) and Code (using this branch) to account for the discrepancies between Code and Prod. I'll do this at some point in the next day or two, so consider it alive for now.