@GHaberis 2017-09-20
Back from the grave.. time to resurrect or finally kill this PR. 

Using the following [blog](https://www.theguardian.com/politics/blog/live/2017/sep/20/keeping-1-public-sector-pay-cap-could-damage-public-services-says-ifs-politics-live) I've ran some tests on code...

**master**
First meaningful paint from lighthouse: 1,060 ms
Consistently Interactive from lighthouse: 11,550 ms
Document weight from dev tools: 527 KB

**gh-faster-liveblog-styles**
First meaningful paint from lighthouse: 930 ms
Consistently Interactive from lighthouse: 10970 ms
Document weight from dev tools: 543 kb

Updated screen grabs below. @SiAdcock the jump in styles is more noticeable in these grabs as they are both taken from code. I actually believe the difference between the before/after screen grabs before was less pronounced as they were taken from my local build which doesn't actually add the styles in the head :sweat_smile: ... :weary:

Before...

![blog-old](https://user-images.githubusercontent.com/1590704/30640641-95b55bea-9dfb-11e7-9ef8-eedd2373791f.gif)

After...

![blog-new](https://user-images.githubusercontent.com/1590704/30640653-9ea4ba48-9dfb-11e7-9407-f408a5754be9.gif)